08/21

+++ VampireSurvivals/Experience.h
@@ -15,14 +15,14 @@ public:
 	virtual void Update() override;
 	virtual void Render(HDC hdc) override;
 
-	void SetEXP(float exp) { _exp = exp; }
-	float GetEXP() { return _exp; }
+	void SetEXP(int32 exp) { _exp = exp; }
+	int32 GetEXP() { return _exp; }
 
 	//virtual void OnComponentBeginOverlap(Collider* collider, Collider* other) override;
 	//virtual void OnComponentEndOverlap(Collider* collider, Collider* other) override;
 
 protected:
 	vector<Sprite*> _spriteEXP[3] = {};
-	float _exp;
+	int32 _exp;
 };
